Light_Set, Light_Pool, Lane_Set, and Lane_Pool
I wrote these four classes. The main purpose for writing this class is that I want to light and car to increment simultaneously but separate from each other (instead of light 1 lane 1 light 2 lane 2 in a unit second, I want it to be light 1 light 2 lane 1 lane 2). This is just a more accurate model. You can see that Pool classes are similar, and Lane classes are similar. I do it this way because I want the structure of the map to be manipulable. Please check and let me know if you have any question. TCSConstant should be changed though. I need to have more accurate data.