Finally, I think I fix the intersection problem
In this commit, I manage to find a way to account for the intersection problem we discussed last class. Please check Lane_Pool class for more detail, but generally, you have a class named VehiclePool, which will account for cars both in a lane or not. It will have many connection with the Lane_Pool class.