Changed the way speed_limit is assigned so it is consistent with other variables in NormalCar
I added a function to generate the speed limit based on guassian distribution just like the other random variables. btw you prob know this but if you need to make the random variables more varied or less varied, you can simply mess with the denominator used when calculating them in the method. For example: in genRandSpeedLimit(), increasing the 9.0 in the equation:
speed_limit = rand.nextGaussian() * AUTOMATEDFINALVELOCITY / 9.0 + AUTOMATEDFINALVELOCITY;
to 10.0 would cause the values to be less varied and closer to the AUTOMATEDFINALVELOCITY value.