Finished turn safety angle variable
Made functions to calculate/generate the safety angle for cars while turning. Also new function for getting the angle between two turning cars. If the front car has completed its turn, the distance will be set to 0 so the back car will be allowed to complete its turn as if the front car wasn't there.
Also, once a car completes its final turn increment, it is treated as traveling straight, and the safety distance is automatically updated based on its speed coming out of the turn.